pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dbx_schema"]

[project]
name = "dbx-schema"
version = "0.1.0"
description = "Databricks Unity Catalog schema extractor"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "databricks-sdk>=0.91.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.12",
]

[project.scripts]
dbx-schema = "dbx_schema.cli:app"

[dependency-groups]
dev = ["pytest>=8.0", "pytest-mock>=3.14", "ruff>=0.9"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
